feat(ios): add onboarding_started event and extract OnboardingStepView
Analytics:
- New `onboarding_started` event fires once per session when the user
  enters the multi-step questionnaire — email fresh path (welcome→next
  transition) or social fresh path (post-OAuth entry). Matches the web
  funnel's `onboarding_started` and fixes the canonical iOS funnel order
  broken by the step reorder (step_completed used to precede signup_started).
- Social users now fire `onboarding_started` on entry instead of
  `onboarding_resumed`. Semantically correct — they're starting, not
  resuming. Social dashboards counting resumed events need to switch to
  started.
- Legacy `wasEmailRegistered` fallback now wipes persisted storage when
  `validateSession()` fails, preventing the infinite welcome→mid-step
  loop on cold starts after session expiry.

Refactor:
- Extracted `OnboardingStepView` + `ScrollMetrics` to their own file
  (OnboardingStepView.swift). OnboardingFlow.swift was at 513 lines
  after the analytics additions — past the 500-line SwiftLint limit.
  Zero behaviour change, pure split.

New canonical iOS funnel:
app_opened → welcome_screen_viewed → onboarding_started →
onboarding_step_completed × N → signup_started → signup_completed →
pin_setup_completed → first_budget_created
